Draft refactoring changes
2 tests still failing
