+++ breadbox/breadbox/service/dataset.py
@@ -1,14 +1,21 @@
+import json
 import logging
+import pandas as pd
+from typing import Any, Dict, Optional
 
+from breadbox.crud.access_control import user_has_access_to_group
 from breadbox.crud import dataset as dataset_crud
 from breadbox.db.session import SessionWithUser
+from breadbox.io.data_validation import annotation_type_to_pandas_column_type
 from breadbox.io.filestore_crud import get_slice
-from breadbox.models.dataset import MatrixDataset
+from breadbox.models.dataset import AnnotationType, MatrixDataset, TabularDataset
 from breadbox.schemas.dataset import (
+    ColumnMetadata,
     FeatureSampleIdentifier,
     MatrixDimensionsInfo,
+    TabularDimensionsInfo,
 )
-from breadbox.schemas.custom_http_exception import UserError
+from breadbox.schemas.custom_http_exception import UserError, DatasetAccessError
 from breadbox.service import metadata as metadata_service
 
 log = logging.getLogger(__name__)
@@ -95,3 +102,157 @@ def get_subsetted_matrix_dataset_df(
         df = df.rename(index=label_by_id)
 
     return df
+
+
+def get_subsetted_tabular_dataset_df(
+    db: SessionWithUser,
+    user: str,
+    dataset: TabularDataset,
+    tabular_dimensions_info: TabularDimensionsInfo,
+    strict: bool,
+) -> pd.DataFrame:
+    """
+    Load a dataframe containing data for the specified indices and columns.
+    If the indices are specified by label, then return a result indexed by labels
+    If either indices or columns are not specified, return all indices or columns
+    By default, if indices and identifier not specified, then dimension ids are used as identifier
+    """
+    if not user_has_access_to_group(dataset.group, user, write_access=True):
+        raise DatasetAccessError(f"User {user} does not have access to dataset")
+
+    # If labels were given as the filter, get the corresponding set of given ids
+    index_filtered_by_label = (
+        tabular_dimensions_info.identifier == FeatureSampleIdentifier.label
+    ) and tabular_dimensions_info.indices is not None
+    if index_filtered_by_label:
+        all_dataset_labels_by_id = metadata_service.get_tabular_dataset_labels_by_id(
+            db, dataset
+        )
+        user_specified_given_ids = {}
+        all_dataset_ids_by_label = {
+            label: id for id, label in all_dataset_labels_by_id.items()
+        }
+        user_specified_given_ids = [
+            all_dataset_ids_by_label[label] for label in tabular_dimensions_info.indices
+        ]
+    elif tabular_dimensions_info.indices is None:
+        user_specified_given_ids = None
+    else:
+        user_specified_given_ids = tabular_dimensions_info.indices
+
+    tabular_subset_df = dataset_crud.get_subset_of_tabular_data_as_df(
+        db=db,
+        dataset=dataset,
+        column_names=tabular_dimensions_info.columns,
+        index_given_ids=user_specified_given_ids,
+    )  # TODO: is this still a "pivot df" somehow?
+
+    if index_filtered_by_label:
+        # Rename the resulting column with dimension ids to their labels
+        tabular_subset_df = tabular_subset_df.replace(
+            {"dimension_given_id": all_dataset_labels_by_id}
+        )
+
+    # If 'strict' raise error
+    missing_columns, missing_indices = _get_missing_tabular_columns_and_indices(
+        tabular_subset_df,
+        tabular_dimensions_info.columns,
+        tabular_dimensions_info.indices,
+        dataset.id,
+    )
+    if strict and (missing_columns or missing_indices):
+        raise UserError(msg=_get_truncated_message(missing_columns, missing_indices))
+
+    # If df is empty, there is no 'value' key to index by
+    if tabular_subset_df.empty:
+        return tabular_subset_df
+
+    # Need to index by "value" after checking if empty db bc empty db has no 'value' keyword
+    subsetted_tabular_dataset_df = tabular_subset_df["value"]
+    # set typing for columns
+    col_dtypes = _get_column_types(
+        dataset.columns_metadata, tabular_dimensions_info.columns
+    )
+    subsetted_tabular_dataset_df = _convert_subsetted_tabular_df_dtypes(
+        subsetted_tabular_dataset_df, col_dtypes, dataset.columns_metadata
+    )
+    return subsetted_tabular_dataset_df
+
+
+def _convert_subsetted_tabular_df_dtypes(
+    df: pd.DataFrame,
+    dtype_map: Dict[str, Any],
+    dataset_columns_metadata: Dict[str, ColumnMetadata],
+):
+    # Replace string boolean values with boolean
+    for col, dtype in dtype_map.items():
+        column = df[col]
+        if dtype == pd.BooleanDtype():
+            column = column.replace({"True": True, "False": False})
+        column = column.astype(dtype)
+        # NOTE: if col type is list string, convert to list. col dtype will be changed to object
+        if (
+            dtype == pd.StringDtype()
+            and dataset_columns_metadata[col].col_type == AnnotationType.list_strings
+        ):
+            column = column.apply(lambda x: json.loads(x) if x is not pd.NA else x)
+        df[col] = column
+    return df
+
+
+def _get_missing_tabular_columns_and_indices(
+    df, tabular_columns, tabular_indices, dataset_id
+):
+    missing_columns = set()
+    missing_indices = set()
+    if tabular_columns is not None:
+        found_columns = [x[1] for x in df.columns]
+        missing_columns = set(tabular_columns).difference(found_columns)
+        if len(missing_columns) > 0:
+            log.warning(
+                f"In get_subsetted_tabular_dataset_df, missing columns: {missing_columns} for dataset: {dataset_id}"
+            )
+
+    if tabular_indices is not None:
+        missing_indices = set(tabular_indices).difference(df.index)
+        if len(missing_indices) > 0:
+            log.warning(
+                f"In get_subsetted_tabular_dataset_df, missing indices: {missing_indices} for dataset: {dataset_id}"
+            )
+
+    return missing_columns, missing_indices
+
+
+def _get_truncated_message(missing_tabular_columns, missing_tabular_indices):
+    num_missing_cols = len(missing_tabular_columns)
+    num_missing_indices = len(missing_tabular_indices)
+    shown_missing_cols = (
+        missing_tabular_columns[:20] + ["..."]
+        if num_missing_cols >= 20
+        else missing_tabular_columns
+    )
+    shown_missing_indices = (
+        missing_tabular_indices[:20] + ["..."]
+        if num_missing_indices >= 20
+        else missing_tabular_indices
+    )
+    return f"{num_missing_cols} missing columns: {shown_missing_cols} and {num_missing_indices} missing indices: {shown_missing_indices}"
+
+
+def _get_column_types(columns_metadata, columns: Optional[list[str]]):
+    col_and_column_metadata_pairs = columns_metadata.items()
+    if columns is None:
+        return {
+            col: annotation_type_to_pandas_column_type(column_metadata.col_type)
+            for col, column_metadata in col_and_column_metadata_pairs
+        }
+
+    else:
+        column_types = {}
+        for col, column_metadata in col_and_column_metadata_pairs:
+            if col in columns:
+                column_types[col] = annotation_type_to_pandas_column_type(
+                    column_metadata.col_type
+                )
+
+        return column_types
